+++ components/mapbox/markers/index.ts
@@ -1,3 +1,4 @@
 // 마커 생성 함수들을 중앙에서 export
-export { createTeamMarker } from './TeamMarker';
-export { createCircuitMarker } from './CircuitMarker';
+export { TeamMarkerFactory } from './team/TeamMarkerFactory';
+export { createCircuitMarker } from './circuit/CircuitMarker';
+export { getTeamMarkerConfig, getAllTeamIds } from './team/teamMarkerConfig';